Add tests and refactor!
Tests help with design, for sure.

Refactored the GameState methods to static methods which allows to
inject a GameState for tests.

Moved the client's guess generation out of the model since it doesn't
belong there and will be replaced when we add actual clients anyways.

Added some simple unit tests which will be picked up by Github actions
from now on!